requirements.txt
numpy==1.24.1
Pillow==10.4.0
streamlit==1.31.1
tensorflow==2.17.0
tensorflow_intel==2.17.0
